Consolidate IV tests into test_signals_iv.py, delete test_iv_signals.py
